refactor: add useFareProductColor for school ticket color support
